refactor: accurate parameter names / createTask struct
- The api query parameters will always be the users public id so the
names were changed to reflect this
- Creating a task now involves sending context to allow polymorphic
assignment and as such is represented in sending the data in the body of
the requests when creating a new task

NOTE:
To create a task at this point means to create a default task with 3
pieces of context:
 - context type (meaning 'project', 'course', 'personal')
 - context id (the id of the related type as above)
 - parent task id (should the task be created inside another task this
will have point to that tasks id)

All of these fields are nullable, such that a task may be 'unlinked' to
anything.

If a task being created has a context type it WILL have a context id and
vice versa.